+++ sdk/lib/util/index.ts
@@ -1,5 +1,3 @@
-import * as T from "../types"
-
 import "./nullIfEmpty"
 import "./fileHelper"
 import "../store/getStore"
@@ -12,26 +10,5 @@ export { GetServiceInterface, getServiceInterface } from "./getServiceInterface"
 export { getServiceInterfaces } from "./getServiceInterfaces"
 export { addressHostToUrl } from "./getServiceInterface"
 export { hostnameInfoToAddress } from "./Hostname"
-// prettier-ignore
-export type FlattenIntersection<T> = 
-T extends ArrayLike<any> ? T :
-T extends object ? {} & {[P in keyof T]: T[P]} :
- T;
-
-export type _<T> = FlattenIntersection<T>
-
-export const isKnownError = (e: unknown): e is T.KnownError =>
-  e instanceof Object && ("error" in e || "error-code" in e)
-
-declare const affine: unique symbol
-
-export type Affine<A> = { [affine]: A }
-
-type NeverPossible = { [affine]: string }
-export type NoAny<A> = NeverPossible extends A
-  ? keyof NeverPossible extends keyof A
-    ? never
-    : A
-  : A
-
+export * from "./typeHelpers"
 export { getDefaultString } from "./getDefaultString"
